Demand & Supply Refactored and Tested. Files Cleaned Up. TOML updated
