+++ message_ix_models/model/water/data/demands.py
@@ -1,19 +1,22 @@
 """Prepare data for adding demands"""
 
 import os
-import builtins
 from collections.abc import Sequence
 from typing import TYPE_CHECKING, Literal, Union
 
 import numpy as np
 import pandas as pd
 import xarray as xr
-from message_ix import make_df
+from message_ix import make_df as make_df
+
 
 from message_ix_models.util import broadcast, minimum_version, package_data_path
 
+from message_ix_models.model.water.data.demand_rules import *
+
 if TYPE_CHECKING:
     from message_ix_models import Context
+
 
 
 def get_basin_sizes(
@@ -21,10 +24,7 @@ def get_basin_sizes(
 ) -> Sequence[Union[pd.Series, Literal[0]]]:
     """Returns the sizes of developing and developed basins for a given node"""
     temp = basin[basin["BCU_name"] == node]
-    print(temp)
     sizes = temp.pivot_table(index=["STATUS"], aggfunc="size")
-    print(sizes)
-    # sizes_### = sizes["###"] if "###" in sizes.index else 0
     sizes_dev = sizes["DEV"] if "DEV" in sizes.index else 0
     sizes_ind = sizes["IND"] if "IND" in sizes.index else 0
     return_tuple: tuple[Union[pd.Series, Literal[0]], Union[pd.Series, Literal[0]]] = (
@@ -47,119 +47,108 @@ def set_target_rate(df: pd.DataFrame, node: str, year: int, target: float) -> No
             df.at[index, "value"] = target
 
 
-def set_target_rate_developed(df: pd.DataFrame, node: str, target: float) -> None:
-    """Sets target rate for a developed basin"""
-    set_target_rate(df, node, 2030, target)
-
-
-def set_target_rate_developing(df: pd.DataFrame, node: str, target: float) -> None:
-    """Sets target rate for a developing basin"""
-    for i in df.index:
-        if df.at[i, "node"] == node and df.at[i, "year"] == 2030:
-            value_2030 = df.at[i, "value"]
-            break
-
-    set_target_rate(
-        df,
-        node,
-        2035,
-        (value_2030 + target) / 2,
-    )
-    set_target_rate(df, node, 2040, target)
-
-
-def set_target_rates(df: pd.DataFrame, basin: pd.DataFrame, val: float) -> None:
-    """Sets target rates for all nodes in a given basin"""
-    for node in df.node.unique():
-        dev_size, ind_size = get_basin_sizes(basin, node)
-        if dev_size >= ind_size:
-            set_target_rate_developed(df, node, val)
-        else:
-            set_target_rate_developing(df, node, val)
-
-
 def target_rate(df: pd.DataFrame, basin: pd.DataFrame, val: float) -> pd.DataFrame:
     """
-    Sets target connection and sanitation rates for SDG scenario.
-    The function filters out the basins as developing and
-    developed based on the countries overlapping basins.
-    If the number of developing countries in the basins are
-    more than basin is categorized as developing and vice versa.
-    If the number of developing and developed countries are equal
-    in a basin, then the basin is assumed developing.
-    For developed basins, target is set at 2030.
-    For developing basins, the access target is set at
-    2040 and 2035 target is the average of
-    2030 original rate and 2040 target.
-
-    Returns
-    -------
-        df (pandas.DataFrame): Data frame with updated value column.
+    Sets target rates for all nodes in a given basin
     """
-    set_target_rates(df, basin, val)
-    return df
+    for node in df.node.unique():
+        dev_size, ind_size = get_basin_sizes(basin, node)
+
+        is_developed = dev_size >= ind_size
+        match is_developed:
+            case True:
+                set_target_rate(df, node, 2030, val)
+            case False:
+                for i in df.index:
+                    if df.at[i, "node"] == node and df.at[i, "year"] == 2030:
+                        value_2030 = df.at[i, "value"]
+                        break
+                set_target_rate(df, node, 2035, (value_2030 + val) / 2)
+                set_target_rate(df, node, 2040, val)
 
 
 def target_rate_trt(df: pd.DataFrame, basin: pd.DataFrame) -> pd.DataFrame:
     """
     Sets target treatment rates for SDG scenario. The target value for
-    developed and developing region is making sure that the amount of untreated
+    developed and developing regions is making sure that the amount of untreated
     wastewater is halved beyond 2030 & 2040 respectively.
 
     Returns
     -------
     data : pandas.DataFrame
     """
-
-    value = []
-    for i in df.node.unique():
-        temp = basin[basin["BCU_name"] == i]
-
-        sizes = temp.pivot_table(index=["STATUS"], aggfunc="size")
-
-        if len(sizes) > 1:
-            if sizes["DEV"] > sizes["IND"] or sizes["DEV"] == sizes["IND"]:
-                for j in df[df["node"] == i][df[df["node"] == i]["year"] >= 2040].index:
-                    temp = df[df["node"] == i][df[df["node"] == i]["year"] >= 2040].at[
-                        j, "value"
-                    ]
-                    temp = temp + (1 - temp) / 2
-                    value.append([j, np.float64(temp)])
-            else:
-                for j in df[df["node"] == i][df[df["node"] == i]["year"] >= 2030].index:
-                    temp = df[df["node"] == i][df[df["node"] == i]["year"] >= 2030].at[
-                        j, "value"
-                    ]
-                    temp = temp + (1 - temp) / 2
-                    value.append([j, np.float64(temp)])
-        else:
-            if sizes.index[0] == "DEV":
-                for j in df[df["node"] == i][df[df["node"] == i]["year"] >= 2040].index:
-                    temp = df[df["node"] == i][df[df["node"] == i]["year"] >= 2040].at[
-                        j, "value"
-                    ]
-                    temp = temp + (1 - temp) / 2
-                    value.append([j, np.float64(temp)])
-            else:
-                for j in df[df["node"] == i][df[df["node"] == i]["year"] >= 2030].index:
-                    temp = df[df["node"] == i][df[df["node"] == i]["year"] >= 2030].at[
-                        j, "value"
-                    ]
-                    temp = temp + (1 - temp) / 2
-                    value.append([j, np.float64(temp)])
-
-    valuetest = pd.DataFrame(data=value, columns=["Index", "Value"])
-
-    for i in range(len(valuetest["Index"])):
-        df.at[valuetest["Index"][i], "Value"] = valuetest["Value"][i]
-
+    updates = []  # Will hold tuples of (index, new_value)
+    
+    for node in df.node.unique():
+        basin_node = basin[basin["BCU_name"] == node]
+        sizes = basin_node.pivot_table(index=["STATUS"], aggfunc="size")
+        
+        # Use pattern matching to decide on the threshold year.
+        is_dev = sizes["DEV"] >= sizes["IND"]
+        match len(sizes):
+            case n if n > 1:
+                match is_dev:
+                    case True:
+                        threshold = 2040
+                    case False:
+                        threshold = 2030
+            case 1:
+                match sizes.index[0]:
+                    case "DEV":
+                        threshold = 2040
+                    case _:
+                        threshold = 2030
+        
+        # Filter rows for this node and the chosen threshold year.
+        node_rows = df[(df["node"] == node) & (df["year"] >= threshold)]
+        for j in node_rows.index:
+            old_val = df.at[j, "value"]
+            new_val = old_val + (1 - old_val) / 2
+            updates.append((j, np.float64(new_val)))
+    
+    # Create a temporary DataFrame from the updates.
+    update_df = pd.DataFrame(updates, columns=["Index", "Value"])
+    
+    # Update the main DataFrame with new values.
+    for _, row in update_df.iterrows():
+        df.at[row["Index"], "Value"] = row["Value"]
+    
+    # Combine new values with original ones.
     real_value = df["Value"].combine_first(df["value"])
-
     df.drop(["value", "Value"], axis=1, inplace=True)
-
     df["value"] = real_value
+    
     return df
 
+def _preprocess_availability_data(df: pd.DataFrame, monthly: bool = False, df_x: pd.DataFrame = None, info = None) -> pd.DataFrame:
+    """
+    Preprocesses availability data
+
+    Parameters
+    ----------
+    df : pd.DataFrame
+    monthly : bool
+    df_x : pd.DataFrame
+    info : Context
+
+    Returns
+    -------
+    df : pd.DataFrame
+    """
+    df.drop(["Unnamed: 0"], axis=1, inplace=True)
+    df.index = df_x["BCU_name"].index
+    df = df.stack().reset_index()
+    df.columns = pd.Index(["Region", "years", "value"])
+    df.fillna(0, inplace=True)
+    df.reset_index(drop=True, inplace=True)
+    df["year"] = pd.DatetimeIndex(df["years"]).year
+    df["time"] = "year" if not monthly else pd.DatetimeIndex(df["years"]).month
+    df["Region"] = df["Region"].map(df_x["BCU_name"])
+    df2210 = df[df["year"] == 2100].copy()
+    df2210["year"] = 2110
+    df = pd.concat([df, df2210])
+    df = df[df["year"].isin(info.Y)]
+    return df
 
 def read_water_availability(context: "Context") -> Sequence[pd.DataFrame]:
     """
@@ -183,108 +172,43 @@ def read_water_availability(context: "Context") -> Sequence[pd.DataFrame]:
         "water", "delineation", f"basins_by_region_simpl_{context.regions}.csv"
     )
     df_x = pd.read_csv(PATH)
+    monthly = False
+    match context.time:
+        case "year":
+            # path for reading basin delineation file
+            path1 = package_data_path(
+                "water",
+                "availability",
+                f"qtot_5y_{context.RCP}_{context.REL}_{context.regions}.csv",
+            )
+            # Reading data, the data is spatially and temprally aggregated from GHMs
+            path2 = package_data_path(
+                "water",
+                "availability",
+                f"qr_5y_{context.RCP}_{context.REL}_{context.regions}.csv",
+            )
+        case "month":
+            monthly = True
+            path1 = package_data_path(
+                "water",
+                "availability",
+                f"qtot_5y_m_{context.RCP}_{context.REL}_{context.regions}.csv",
+            )
 
-    if "year" in context.time:
-        # path for reading basin delineation file
-        PATH = package_data_path(
-            "water", "delineation", f"basins_by_region_simpl_{context.regions}.csv"
-        )
-        df_x = pd.read_csv(PATH)
-        # Adding freshwater supply constraints
-        # Reading data, the data is spatially and temprally aggregated from GHMs
-        path1 = package_data_path(
-            "water",
-            "availability",
-            f"qtot_5y_{context.RCP}_{context.REL}_{context.regions}.csv",
-        )
-        # Read rcp 2.6 data
-        df_sw = pd.read_csv(path1)
-        df_sw.drop(["Unnamed: 0"], axis=1, inplace=True)
+            # Reading data, the data is spatially and temporally aggregated from GHMs
+            path2 = package_data_path(
+                "water",
+                "availability",
+                f"qr_5y_m_{context.RCP}_{context.REL}_{context.regions}.csv",
+            )
+        case _:
+            raise ValueError(f"Invalid time period: {context.time}")
 
-        df_sw.index = df_x["BCU_name"].index
-        df_sw = df_sw.stack().reset_index()
-        df_sw.columns = pd.Index(["Region", "years", "value"])
-        df_sw.fillna(0, inplace=True)
-        df_sw.reset_index(drop=True, inplace=True)
-        df_sw["year"] = pd.DatetimeIndex(df_sw["years"]).year
-        df_sw["time"] = "year"
-        df_sw["Region"] = df_sw["Region"].map(df_x["BCU_name"])
-        df_sw2210 = df_sw[df_sw["year"] == 2100].copy()
-        df_sw2210["year"] = 2110
-        df_sw = pd.concat([df_sw, df_sw2210])
-        df_sw = df_sw[df_sw["year"].isin(info.Y)]
+    df_sw = pd.read_csv(path1)
+    df_sw = _preprocess_availability_data(df_sw, monthly=monthly, df_x = df_x, info = info)
 
-        # Adding groundwater supply constraints
-        # Reading data, the data is spatially and temprally aggregated from GHMs
-        path1 = package_data_path(
-            "water",
-            "availability",
-            f"qr_5y_{context.RCP}_{context.REL}_{context.regions}.csv",
-        )
-
-        # Read groundwater data
-        df_gw = pd.read_csv(path1)
-        df_gw.drop(["Unnamed: 0"], axis=1, inplace=True)
-        df_gw.index = df_x["BCU_name"].index
-        df_gw = df_gw.stack().reset_index()
-        df_gw.columns = pd.Index(["Region", "years", "value"])
-        df_gw.fillna(0, inplace=True)
-        df_gw.reset_index(drop=True, inplace=True)
-        df_gw["year"] = pd.DatetimeIndex(df_gw["years"]).year
-        df_gw["time"] = "year"
-        df_gw["Region"] = df_gw["Region"].map(df_x["BCU_name"])
-        df_gw2210 = df_gw[df_gw["year"] == 2100].copy()
-        df_gw2210["year"] = 2110
-        df_gw = pd.concat([df_gw, df_gw2210])
-        df_gw = df_gw[df_gw["year"].isin(info.Y)]
-
-    else:
-        # Adding freshwater supply constraints
-        # Reading data, the data is spatially and temprally aggregated from GHMs
-        path1 = package_data_path(
-            "water",
-            "availability",
-            f"qtot_5y_m_{context.RCP}_{context.REL}_{context.regions}.csv",
-        )
-        df_sw = pd.read_csv(path1)
-        df_sw.drop(["Unnamed: 0"], axis=1, inplace=True)
-
-        df_sw.index = df_x["BCU_name"].index
-        df_sw = df_sw.stack().reset_index()
-        df_sw.columns = pd.Index(["Region", "years", "value"])
-        df_sw.sort_values(["Region", "years", "value"], inplace=True)
-        df_sw.fillna(0, inplace=True)
-        df_sw.reset_index(drop=True, inplace=True)
-        df_sw["year"] = pd.DatetimeIndex(df_sw["years"]).year
-        df_sw["time"] = pd.DatetimeIndex(df_sw["years"]).month
-        df_sw["Region"] = df_sw["Region"].map(df_x["BCU_name"])
-        df_sw2210 = df_sw[df_sw["year"] == 2100].copy()
-        df_sw2210["year"] = 2110
-        df_sw = pd.concat([df_sw, df_sw2210])
-        df_sw = df_sw[df_sw["year"].isin(info.Y)]
-
-        # Reading data, the data is spatially and temporally aggregated from GHMs
-        path1 = package_data_path(
-            "water",
-            "availability",
-            f"qr_5y_m_{context.RCP}_{context.REL}_{context.regions}.csv",
-        )
-        df_gw = pd.read_csv(path1)
-        df_gw.drop(["Unnamed: 0"], axis=1, inplace=True)
-
-        df_gw.index = df_x["BCU_name"].index
-        df_gw = df_gw.stack().reset_index()
-        df_gw.columns = pd.Index(["Region", "years", "value"])
-        df_gw.sort_values(["Region", "years", "value"], inplace=True)
-        df_gw.fillna(0, inplace=True)
-        df_gw.reset_index(drop=True, inplace=True)
-        df_gw["year"] = pd.DatetimeIndex(df_gw["years"]).year
-        df_gw["time"] = pd.DatetimeIndex(df_gw["years"]).month
-        df_gw["Region"] = df_gw["Region"].map(df_x["BCU_name"])
-        df_gw2210 = df_gw[df_gw["year"] == 2100].copy()
-        df_gw2210["year"] = 2110
-        df_gw = pd.concat([df_gw, df_gw2210])
-        df_gw = df_gw[df_gw["year"].isin(info.Y)]
+    df_gw = pd.read_csv(path2)
+    df_gw = _preprocess_availability_data(df_gw, monthly=monthly, df_x = df_x, info = info)
 
     return df_sw, df_gw
 
@@ -310,49 +234,52 @@ def add_water_availability(context: "Context") -> dict[str, pd.DataFrame]:
     # Reading data, the data is spatially and temprally aggregated from GHMs
 
     df_sw, df_gw = read_water_availability(context)
+    water_availability = []
+    for rule in WATER_AVAILABILITY:
+        match rule["df_source"]:
+            case "df_sw":
+                df_source = df_sw
+            case "df_gw":
+                df_source = df_gw
+            case _:
+                raise ValueError(f"Invalid df_source: {rule['df_source']}")
 
-    dmd_df = make_df(
-        "demand",
-        node="B" + df_sw["Region"].astype(str),
-        commodity="surfacewater_basin",
-        level="water_avail_basin",
-        year=df_sw["year"],
-        time=df_sw["time"],
-        value=-df_sw["value"],
-        unit="km3/year",
-    )
+        dmd_df = make_df(
+            rule["type"],
+            node="B"+ eval_field(rule["node"], df_source),
+            commodity=rule["commodity"],
+            level=rule["level"],
+            year=eval_field(rule["year"], df_source),
+            time=eval_field(rule["time"], df_source),
+            value= -eval_field(rule["value"], df_source),
+            unit=rule["unit"],
+        )
+        water_availability.append(dmd_df)
 
-    dmd_df = pd.concat(
-        [
-            dmd_df,
-            make_df(
-                "demand",
-                node="B" + df_gw["Region"].astype(str),
-                commodity="groundwater_basin",
-                level="water_avail_basin",
-                year=df_gw["year"],
-                time=df_gw["time"],
-                value=-df_gw["value"],
-                unit="km3/year",
-            ),
-        ]
-    )
+    dmd_df = pd.concat(water_availability)
 
     dmd_df["value"] = dmd_df["value"].apply(lambda x: x if x <= 0 else 0)
 
     results["demand"] = dmd_df
 
-    # share constraint lower bound on groundwater
-    df_share = make_df(
-        "share_commodity_lo",
-        shares="share_low_lim_GWat",
-        node_share="B" + df_gw["Region"].astype(str),
-        year_act=df_gw["year"],
-        time=df_gw["time"],
-        value=df_gw["value"]
-        / (df_sw["value"] + df_gw["value"])
-        * 0.95,  # 0.95 buffer factor to avoid numerical error
-        unit="-",
+
+    share_constraints_gw = []
+    for rule in SHARE_CONSTRAINTS_GW:
+        # share constraint lower bound on groundwater
+        if rule["df_source1"] == "df_gw" and rule["df_source2"] == "df_sw":
+            df_source1 = df_gw
+            df_source2 = df_sw
+        else:
+            raise ValueError(f"Invalid df_source: {rule['df_source1']} or {rule['df_source2']}")
+        
+        df_share = make_df(
+            rule["type"],
+            shares=rule["shares"],
+            node_share="B" + eval_field(rule["node"], df_source1),
+            year_act=eval_field(rule["year"], df_source1),
+            time=eval_field(rule["time"], df_source1),
+            value=eval_field(rule["value"], df_source1, df_source2),             
+            unit=rule["unit"],
     )
 
     df_share["value"] = df_share["value"].fillna(0)
@@ -404,8 +331,9 @@ def add_irrigation_demand(context: "Context") -> dict[str, pd.DataFrame]:
 
     return results
 
+    
 
-def _preprocess_demand_data(context: "Context") -> pd.DataFrame:
+def _preprocess_demand_data_stage1(context: "Context") -> pd.DataFrame:
     # read and clean raw demand data to standardized format
     region = f"{context.regions}"
     # get data path using package_data_path
@@ -455,6 +383,34 @@ def _preprocess_demand_data(context: "Context") -> pd.DataFrame:
         df_dmds = pd.concat([df_dmds, df_m])
     return df_dmds
 
+def _preprocess_demand_data_stage2(df_dmds: pd.DataFrame) -> dict[str, pd.DataFrame]:
+    """
+    Pre-process the DataFrame to prepare it for the rule evaluation.
+    """
+    variables_operations = {
+        "urban_withdrawal2_baseline": {"df_name": "urban_withdrawal_df", "reset_index": False},
+        "rural_withdrawal_baseline": {"df_name": "rural_withdrawal_df", "reset_index": False},
+        "manufacturing_withdrawal_baseline": {"df_name": "industrial_withdrawals_df", "reset_index": False},
+        "manufacturing_return_baseline": {"df_name": "industrial_return_df", "reset_index": False},
+        "urban_return2_baseline": {"df_name": "urban_return_df", "reset_index": True},
+        "rural_return_baseline": {"df_name": "rural_return_df", "reset_index": True},
+        "urban_connection_rate_baseline": {"df_name": "urban_connection_rate_df", "reset_index": True},
+        "rural_connection_rate_baseline": {"df_name": "rural_connection_rate_df", "reset_index": True},
+        "urban_treatment_rate_baseline": {"df_name": "urban_treatment_rate_df", "reset_index": True},
+        "rural_treatment_rate_baseline": {"df_name": "rural_treatment_rate_df", "reset_index": True},
+        "urban_recycling_rate_baseline": {"df_name": "df_recycling", "reset_index": True},
+    }
+    Results = {}
+    for variable, attrs in variables_operations.items():
+        df_name, reset_index = attrs["df_name"], attrs["reset_index"]
+        df_name = df_dmds[df_dmds["variable"] == variable]
+        if reset_index:
+            df_name.reset_index(drop=True, inplace=True)
+        Results[attrs["df_name"]] = df_name
+
+    return Results
+
+
 
 @minimum_version("message_ix 3.7")
 def add_sectoral_demands(context: "Context") -> dict[str, pd.DataFrame]:
@@ -478,29 +434,21 @@ def add_sectoral_demands(context: "Context") -> dict[str, pd.DataFrame]:
     sub_time = context.time
 
     # get standardized input data
-    df_dmds = _preprocess_demand_data(context)
-
-    urban_withdrawal_df = df_dmds[df_dmds["variable"] == "urban_withdrawal2_baseline"]
-    rual_withdrawal_df = df_dmds[df_dmds["variable"] == "rural_withdrawal_baseline"]
-    industrial_withdrawals_df = df_dmds[df_dmds["variable"] == "manufacturing_withdrawal_baseline"]
-    industrial_return_df = df_dmds[df_dmds["variable"] == "manufacturing_return_baseline"]
-    urban_return_df = df_dmds[df_dmds["variable"] == "urban_return2_baseline"]
-    urban_return_df.reset_index(drop=True, inplace=True)
-    rural_return_df = df_dmds[df_dmds["variable"] == "rural_return_baseline"]
-    rural_return_df.reset_index(drop=True, inplace=True)
-    urban_connection_rate_df = df_dmds[df_dmds["variable"] == "urban_connection_rate_baseline"]
-    urban_connection_rate_df.reset_index(drop=True, inplace=True)
-    rural_connection_rate_df = df_dmds[df_dmds["variable"] == "rural_connection_rate_baseline"]
-    rural_connection_rate_df.reset_index(drop=True, inplace=True)
-
-    urban_treatment_rate_df = df_dmds[df_dmds["variable"] == "urban_treatment_rate_baseline"]
-    urban_treatment_rate_df.reset_index(drop=True, inplace=True)
-
-    rural_treatment_rate_df = df_dmds[df_dmds["variable"] == "rural_treatment_rate_baseline"]
-    rural_treatment_rate_df.reset_index(drop=True, inplace=True)
-
-    df_recycling = df_dmds[df_dmds["variable"] == "urban_recycling_rate_baseline"]
-    df_recycling.reset_index(drop=True, inplace=True)
+    df_dmds = _preprocess_demand_data_stage1(context)
+    
+    # Process data and unpack results
+    processed_data = _preprocess_demand_data_stage2(df_dmds)
+    urban_withdrawal_df = processed_data["urban_withdrawal_df"]
+    rural_withdrawal_df = processed_data["rural_withdrawal_df"]
+    industrial_withdrawals_df = processed_data["industrial_withdrawals_df"]
+    industrial_return_df = processed_data["industrial_return_df"]
+    urban_return_df = processed_data["urban_return_df"]
+    rural_return_df = processed_data["rural_return_df"]
+    urban_connection_rate_df = processed_data["urban_connection_rate_df"]
+    rural_connection_rate_df = processed_data["rural_connection_rate_df"]
+    urban_treatment_rate_df = processed_data["urban_treatment_rate_df"]
+    rural_treatment_rate_df = processed_data["rural_treatment_rate_df"]
+    df_recycling = processed_data["df_recycling"]
 
     if context.SDG != "baseline":
         if context.SDG == "SDG":
@@ -566,226 +514,140 @@ def add_sectoral_demands(context: "Context") -> dict[str, pd.DataFrame]:
         all_rates.to_csv(save_path / "all_rates_SSP2.csv", index=False)
 
     # urban water demand and return. 1e-3 from mcm to km3
-    urban_mw = urban_withdrawal_df.reset_index(drop=True)
-    urban_mw = urban_mw.merge(
-        urban_connection_rate_df.drop(columns=["variable", "time"]).rename(columns={"value": "rate"})
-    )
-    urban_mw["value"] = (1e-3 * urban_mw["value"]) * urban_mw["rate"]
+    urban_dmds = []
+    for r in URBAN_DEMAND:
+        match r["commodity"]:
+            case "urban_mw":
+                urban_mw = urban_withdrawal_df.reset_index(drop=True)
+                urban_mw = urban_mw.merge(
+                    urban_connection_rate_df.drop(columns=["variable", "time"]).rename(columns={"value": "rate"})
+                )
+                urban_mw["value"] = (1e-3 * urban_mw["value"]) * urban_mw["rate"]
 
-    dmd_df = make_df(
-        "demand",
-        node="B" + urban_mw["node"],
-        commodity="urban_mw",
-        level="final",
-        year=urban_mw["year"],
-        time=urban_mw["time"],
-        value=urban_mw["value"],
-        unit="km3/year",
-    )
-    urban_dis = urban_withdrawal_df.reset_index(drop=True)
-    urban_dis = urban_dis.merge(
-        urban_connection_rate_df.drop(columns=["variable", "time"]).rename(columns={"value": "rate"})
-    )
-    urban_dis["value"] = (1e-3 * urban_dis["value"]) * (1 - urban_dis["rate"])
+                urban_dmds.append(load_rules(r, urban_mw))
+            case "urban_disconnected":
+                urban_dis = urban_withdrawal_df.reset_index(drop=True)
+                urban_dis = urban_dis.merge(
+                    urban_connection_rate_df.drop(columns=["variable", "time"]).rename(columns={"value": "rate"})
+                )
+                urban_dis["value"] = (1e-3 * urban_dis["value"]) * (1 - urban_dis["rate"])
 
-    dmd_df = pd.concat(
-        [
-            dmd_df,
-            make_df(
-                "demand",
-                node="B" + urban_dis["node"],
-                commodity="urban_disconnected",
-                level="final",
-                year=urban_dis["year"],
-                time=urban_dis["time"],
-                value=urban_dis["value"],
-                unit="km3/year",
-            ),
-        ]
-    )
+                urban_dmds.append(load_rules(r, urban_dis))
+            case _:
+                raise ValueError(f"Invalid commodity: {r['commodity']}")
+    urban_dmds = pd.concat(urban_dmds)
+    dmd_df = urban_dmds
     # rural water demand and return
-    rural_mw = rual_withdrawal_df.reset_index(drop=True)
-    rural_mw = rural_mw.merge(
-        rural_connection_rate_df.drop(columns=["variable", "time"]).rename(columns={"value": "rate"})
-    )
-    rural_mw["value"] = (1e-3 * rural_mw["value"]) * rural_mw["rate"]
+    rural_dmds = []
+    for r in RURAL_DEMAND:
+        match r["commodity"]:
+            case "rural_mw":
+                rural_mw = rural_withdrawal_df.reset_index(drop=True)
+                rural_mw = rural_mw.merge(
+                    rural_connection_rate_df.drop(columns=["variable", "time"]).rename(columns={"value": "rate"})
+                )
+                rural_mw["value"] = (1e-3 * rural_mw["value"]) * rural_mw["rate"]
 
-    dmd_df = pd.concat(
-        [
-            dmd_df,
-            make_df(
-                "demand",
-                node="B" + rural_mw["node"],
-                commodity="rural_mw",
-                level="final",
-                year=rural_mw["year"],
-                time=rural_mw["time"],
-                value=rural_mw["value"],
-                unit="km3/year",
-            ),
-        ]
-    )
+                rural_dmds.append(load_rules(r, rural_mw))
+            case "rural_disconnected":
+                rural_dis = rural_withdrawal_df.reset_index(drop=True)
+                rural_dis = rural_dis.merge(
+                    rural_connection_rate_df.drop(columns=["variable", "time"]).rename(columns={"value": "rate"})
+                )
+                rural_dis["value"] = (1e-3 * rural_dis["value"]) * (1 - rural_dis["rate"])
 
-    rural_dis = rual_withdrawal_df.reset_index(drop=True)
-    rural_dis = rural_dis.merge(
-        rural_connection_rate_df.drop(columns=["variable", "time"]).rename(columns={"value": "rate"})
-    )
-    rural_dis["value"] = (1e-3 * rural_dis["value"]) * (1 - rural_dis["rate"])
+                rural_dmds.append(load_rules(r, rural_dis))
+            case _:
+                raise ValueError(f"Invalid commodity: {r['commodity']}")
 
-    dmd_df = pd.concat(
-        [
-            dmd_df,
-            make_df(
-                "demand",
-                node="B" + rural_dis["node"],
-                commodity="rural_disconnected",
-                level="final",
-                year=rural_dis["year"],
-                time=rural_dis["time"],
-                value=rural_dis["value"],
-                unit="km3/year",
-            ),
-        ]
-    )
+    rural_dmds = pd.concat(rural_dmds)
+    dmd_df = pd.concat([dmd_df, rural_dmds])
 
     # manufactury/ industry water demand and return
-    manuf_mw = industrial_withdrawals_df.reset_index(drop=True)
-    manuf_mw["value"] = 1e-3 * manuf_mw["value"]
+    industrial_dmds = []
+    for r in INDUSTRIAL_DEMAND:
+        match r["commodity"]:
+            case "industry_mw":
+                # manufactury/ industry water demand and return
+                manuf_mw = industrial_withdrawals_df.reset_index(drop=True)
+                manuf_mw["value"] = 1e-3 * manuf_mw["value"]
+                industrial_dmds.append(load_rules(r, manuf_mw))
+                
+            case "industry_uncollected_wst":
+                manuf_uncollected_wst = industrial_return_df.reset_index(drop=True)
+                manuf_uncollected_wst["value"] = 1e-3 * manuf_uncollected_wst["value"]
+                industrial_dmds.append(load_rules(r, manuf_uncollected_wst))
+            case _:
+                raise ValueError(f"Invalid commodity: {r['commodity']}")
+    industrial_dmds = pd.concat(industrial_dmds)
+    dmd_df = pd.concat([dmd_df, industrial_dmds])
 
-    dmd_df = pd.concat(
-        [
-            dmd_df,
-            make_df(
-                "demand",
-                node="B" + manuf_mw["node"],
-                commodity="industry_mw",
-                level="final",
-                year=manuf_mw["year"],
-                time=manuf_mw["time"],
-                value=manuf_mw["value"],
-                unit="km3/year",
-            ),
-        ]
-    )
-
-    manuf_uncollected_wst = industrial_return_df.reset_index(drop=True)
-    manuf_uncollected_wst["value"] = 1e-3 * manuf_uncollected_wst["value"]
-
-    dmd_df = pd.concat(
-        [
-            dmd_df,
-            make_df(
-                "demand",
-                node="B" + manuf_uncollected_wst["node"],
-                commodity="industry_uncollected_wst",
-                level="final",
-                year=manuf_uncollected_wst["year"],
-                time=manuf_uncollected_wst["time"],
-                value=-manuf_uncollected_wst["value"],
-                unit="km3/year",
-            ),
-        ]
-    )
-
-    urban_collected_wst = urban_return_df.reset_index(drop=True)
-    urban_collected_wst = urban_collected_wst.merge(
-        urban_treatment_rate_df.drop(columns=["variable", "time"]).rename(columns={"value": "rate"})
-    )
-    urban_collected_wst["value"] = (
+    # urban collected wastewater
+    urban_collected_wst_df = []
+    for r in URBAN_COLLECTED_WST:
+        urban_collected_wst = urban_return_df.reset_index(drop=True)
+        urban_collected_wst = urban_collected_wst.merge(
+            urban_treatment_rate_df.drop(columns=["variable", "time"]).rename(columns={"value": "rate"})
+        )
+        urban_collected_wst["value"] = (
         1e-3 * urban_collected_wst["value"]
-    ) * urban_collected_wst["rate"]
+        ) * urban_collected_wst["rate"]
+        urban_collected_wst_df.append(load_rules(r, urban_collected_wst))
 
-    dmd_df = pd.concat(
-        [
-            dmd_df,
-            make_df(
-                "demand",
-                node="B" + urban_collected_wst["node"],
-                commodity="urban_collected_wst",
-                level="final",
-                year=urban_collected_wst["year"],
-                time=urban_collected_wst["time"],
-                value=-urban_collected_wst["value"],
-                unit="km3/year",
-            ),
-        ]
-    )
+    urban_collected_wst_df = pd.concat(urban_collected_wst_df)    
+    dmd_df = pd.concat([dmd_df, urban_collected_wst_df])
 
-    rural_collected_wst = rural_return_df.reset_index(drop=True)
-    rural_collected_wst = rural_collected_wst.merge(
-        rural_treatment_rate_df.drop(columns=["variable", "time"]).rename(columns={"value": "rate"})
-    )
-    rural_collected_wst["value"] = (
-        1e-3 * rural_collected_wst["value"]
-    ) * rural_collected_wst["rate"]
+    # rural collected wastewater
+    rural_collected_wst_df = []
+    for r in RURAL_COLLECTED_WST:
+        rural_collected_wst = rural_return_df.reset_index(drop=True)
+        rural_collected_wst = rural_collected_wst.merge(
+            rural_treatment_rate_df.drop(columns=["variable", "time"]).rename(columns={"value": "rate"})
+        )
+        rural_collected_wst["value"] = (
+            1e-3 * rural_collected_wst["value"]
+        ) * rural_collected_wst["rate"]
+        rural_collected_wst_df.append(load_rules(r, rural_collected_wst))
 
-    dmd_df = pd.concat(
-        [
-            dmd_df,
-            make_df(
-                "demand",
-                node="B" + rural_collected_wst["node"],
-                commodity="rural_collected_wst",
-                level="final",
-                year=rural_collected_wst["year"],
-                time=rural_collected_wst["time"],
-                value=-rural_collected_wst["value"],
-                unit="km3/year",
-            ),
-        ]
-    )
-    urban_uncollected_wst = urban_return_df.reset_index(drop=True)
-    urban_uncollected_wst = urban_uncollected_wst.merge(
-        urban_treatment_rate_df.drop(columns=["variable", "time"]).rename(columns={"value": "rate"})
-    )
-    urban_uncollected_wst["value"] = (1e-3 * urban_uncollected_wst["value"]) * (
-        1 - urban_uncollected_wst["rate"]
-    )
+    rural_collected_wst_df = pd.concat(rural_collected_wst_df)
+    dmd_df = pd.concat([dmd_df, rural_collected_wst_df])
 
-    dmd_df = pd.concat(
-        [
-            dmd_df,
-            make_df(
-                "demand",
-                node="B" + urban_uncollected_wst["node"],
-                commodity="urban_uncollected_wst",
-                level="final",
-                year=urban_uncollected_wst["year"],
-                time=urban_uncollected_wst["time"],
-                value=-urban_uncollected_wst["value"],
-                unit="km3/year",
-            ),
-        ]
-    )
+    # urban uncollected wastewater
+    urban_uncollected_wst_df = []
+    for r in URBAN_UNCOLLECTED_WST:
+        urban_uncollected_wst = urban_return_df.reset_index(drop=True)
+        urban_uncollected_wst = urban_uncollected_wst.merge(
+            urban_treatment_rate_df.drop(columns=["variable", "time"]).rename(columns={"value": "rate"})
+        )
+        urban_uncollected_wst["value"] = (1e-3 * urban_uncollected_wst["value"]) * (
+            1 - urban_uncollected_wst["rate"]
+        )
+        urban_uncollected_wst_df.append(load_rules(r, urban_uncollected_wst))
 
-    rural_uncollected_wst = rural_return_df.reset_index(drop=True)
-    rural_uncollected_wst = rural_uncollected_wst.merge(
-        rural_treatment_rate_df.drop(columns=["variable", "time"]).rename(columns={"value": "rate"})
-    )
-    rural_uncollected_wst["value"] = (1e-3 * rural_uncollected_wst["value"]) * (
-        1 - rural_uncollected_wst["rate"]
-    )
+    urban_uncollected_wst_df = pd.concat(urban_uncollected_wst_df)
+    dmd_df = pd.concat([dmd_df, urban_uncollected_wst_df])
 
-    dmd_df = pd.concat(
-        [
-            dmd_df,
-            make_df(
-                "demand",
-                node="B" + rural_uncollected_wst["node"],
-                commodity="rural_uncollected_wst",
-                level="final",
-                year=rural_uncollected_wst["year"],
-                time=rural_uncollected_wst["time"],
-                value=-rural_uncollected_wst["value"],
-                unit="km3/year",
-            ),
-        ]
-    )
+    # rural uncollected wastewater
+    rural_uncollected_wst_df = []
+    for r in RURAL_UNCOLLECTED_WST:
+        rural_uncollected_wst = rural_return_df.reset_index(drop=True)
+        rural_uncollected_wst = rural_uncollected_wst.merge(
+            rural_treatment_rate_df.drop(columns=["variable", "time"]).rename(columns={"value": "rate"})
+        )
+        rural_uncollected_wst["value"] = (1e-3 * rural_uncollected_wst["value"]) * (
+            1 - rural_uncollected_wst["rate"]
+        )
+        rural_uncollected_wst_df.append(load_rules(r, rural_uncollected_wst))
+
+    rural_uncollected_wst_df = pd.concat(rural_uncollected_wst_df)
+    dmd_df = pd.concat([dmd_df, rural_uncollected_wst_df])
+
+
     # Add 2010 & 2015 values as historical activities to corresponding technologies
     h_act = dmd_df[dmd_df["year"].isin([2010, 2015])]
-
     dmd_df = dmd_df[dmd_df["year"].isin(info.Y)]
     results["demand"] = dmd_df
+
 
     # create a list of our conditions
     conditions = [
@@ -819,52 +681,62 @@ def add_sectoral_demands(context: "Context") -> dict[str, pd.DataFrame]:
     h_act["commodity"] = np.select(conditions, values, "Unknown commodity")
     h_act["value"] = h_act["value"].abs()
 
-    hist_act = make_df(
-        "historical_activity",
-        node_loc=h_act["node"],
-        technology=h_act["commodity"],
-        year_act=h_act["year"],
-        mode="M1",
-        time=h_act["time"],
-        value=h_act["value"],
-        unit="km3/year",
-    )
-    results["historical_activity"] = hist_act
+    # Process historical activities using rules
+    historical_activity_df = []
+    for rule in HISTORICAL_ACTIVITY:
+        hist_act = make_df(
+            rule["type"],
+            node_loc=eval_field(rule["node"], h_act),
+            technology=eval_field(rule["technology"], h_act),
+            year_act=eval_field(rule["year"], h_act),
+            mode=rule["mode"],
+            time=eval_field(rule["time"], h_act),
+            value=eval_field(rule["value"], h_act),
+            unit=rule["unit"],
+        )
+        historical_activity_df.append(hist_act)
+    results["historical_activity"] = pd.concat(historical_activity_df)
 
+    # Process historical capacities using rules
     h_cap = h_act[h_act["year"] >= 2015]
     h_cap = (
         h_cap.groupby(["node", "commodity", "level", "year", "unit"])["value"]
         .sum()
         .reset_index()
     )
-
-    hist_cap = make_df(
-        "historical_new_capacity",
-        node_loc=h_cap["node"],
-        technology=h_cap["commodity"],
-        year_vtg=h_cap["year"],
-        value=h_cap["value"] / 5,
-        unit="km3/year",
-    )
-
-    results["historical_new_capacity"] = hist_cap
+    historical_capacity_df = []
+    for rule in HISTORICAL_CAPACITY:
+        hist_cap = make_df(
+            rule["type"],
+            node_loc=eval_field(rule["node"], h_cap),
+            technology=eval_field(rule["technology"], h_cap),
+            year_vtg=eval_field(rule["year"], h_cap),
+            value=eval_field(rule["value"], h_cap),
+            unit=rule["unit"],
+        )
+        historical_capacity_df.append(hist_cap)
+    results["historical_new_capacity"] = pd.concat(historical_capacity_df)
 
     # share constraint lower bound on urban_Water recycling
-    df_share_wat = make_df(
-        "share_commodity_lo",
-        shares="share_wat_recycle",
-        node_share="B" + df_recycling["node"],
-        year_act=df_recycling["year"],
-        value=df_recycling["value"],
-        unit="-",
+    share_constraint_df = []
+    for rule in SHARE_CONSTRAINTS_RECYCLING:
+        df_share_wat = make_df(
+            rule["type"],
+            shares=rule["shares"],
+            node_share="B" + eval_field(rule["node"], df_recycling),
+            year_act=eval_field(rule["year"], df_recycling),
+            value=eval_field(rule["value"], df_recycling),
+            unit=rule["unit"],
     ).pipe(
-        broadcast,
-        time=pd.Series(sub_time),
-    )
+            broadcast,
+            time=pd.Series(sub_time),
+        )
+        share_constraint_df.append(df_share_wat)
+    results["share_commodity_lo"] = pd.concat(share_constraint_df)
+    results["share_commodity_lo"] = results["share_commodity_lo"][results["share_commodity_lo"]["year_act"].isin(info.Y)]
 
-    df_share_wat = df_share_wat[df_share_wat["year_act"].isin(info.Y)]
-    results["share_commodity_lo"] = df_share_wat
 
     return results
+
