+++ message_ix_models/tests/model/water/data/test_demands_parity.py
@@ -6,8 +6,8 @@ import pandas as pd
 from message_ix_models import ScenarioInfo
 from message_ix_models.model.structure import get_codes
 
-from message_ix_models.model.water.data.demands_rf import read_water_availability as new_read_water_availability, add_water_availability as new_add_water_availability, add_sectoral_demands as new_add_sectoral_demands
-from message_ix_models.model.water.data.demands import read_water_availability as old_read_water_availability, add_water_availability as old_add_water_availability, add_sectoral_demands as old_add_sectoral_demands
+from message_ix_models.model.water.data.demands import read_water_availability as new_read_water_availability, add_water_availability as new_add_water_availability, add_sectoral_demands as new_add_sectoral_demands
+from message_ix_models.model.water.data.demands_legacy import read_water_availability as old_read_water_availability, add_water_availability as old_add_water_availability, add_sectoral_demands as old_add_sectoral_demands
 
 import time as pytime
